# src/main/resources/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "critica_backend API"
  description: "critica_backend API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
  - url: "https://seahorse-app-qs7ri.ondigitalocean.app"
  - url: "https://critika.net"
paths:
...
  /api/auth/refresh:
    post:
      tags:
        - "Auth"
        - "User"
      security:
        - AccessToken: [ ]
      description: "Refresh the authentication token"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshToken"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /api/auth/register:
    post:
      tags:
        - "Auth"
        - "User"
      description: "Register a new user"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                type: string
                example: "OK"
  /api/auth/userExists:
    post:
      description: "Check if user exists in the database"
      tags:
        - "Auth"
      requestBody:
        content:
          'application/json':
            schema:
              type: "object"
              properties:
                email:
                  type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: "boolean"
  /api/auth/register-firebase:
    post:
      tags:
        - "Auth"
        - "User"
      security:
        - FirebaseToken: [ ]
      description: "Register a new user with Firebase token"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                type: string
                example: "OK"
  /api/auth/signIn:
    post:
      tags:
        - "Auth"
        - "User"
      security:
        - FirebaseToken: []
      description: "Log in to the application"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignIn"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthToken"
  /api/auth/signout:
    post:
      tags:
        - "Auth"
        - "User"
      security:
        - AccessToken: [ ]
      description: "Log out of the application"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignOut"
        required: true
      responses:
        "204":
          description: "No Content"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoContent"
...
  /api/game/{id}:
    get:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Get game by id"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/game/{id}/addPlayer:
    put:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Add player by name to the game"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/game/{id}/addPlayer/{playerId}:
    put:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Add player to game"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "playerId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/game/{id}/finish:
    post:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Finishes the current game"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/game/{id}/start:
    post:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Starts the game. Note: This endpoint is for testing purposes only. In production, all endpoints for games would be called from the lobby."
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/game/list:
    get:
      tags:
        - "Game"
      security:
        - AccessToken: [ ]
      description: "Get all games. Note: This endpoint is for testing purposes only."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/GameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    FirebaseToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserAuthToken:
      description: "User user token response"
      type: object
      properties:
        token:
          type: string
          example: "JWT_TOKEN"
        refresh_token:
          type: string
          example: "JWT_TOKEN"
    RefreshToken:
      type: object
      properties:
        id:
          type: "string"
        refreshToken:
          type: "string"
    CreateAccount:
      type: object
      properties:
        email:
          type: "string"
          format: "email"
          example: "johndoe@example.com"
        username:
          type: "string"
          pattern: "^[A-Za-z0-9]*$"
          example: "asfgasfgasfg"
        password:
          type: "string"
          pattern: "^[A-Za-z0-9!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?]{8,}$"
          example: "p@ssw0rd123!"
        playerName:
          type: "string"
          pattern: "^[A-Za-zА-Яа-я0-9]+(?: [A-Za-zА-Яа-я0-9]+)*$"
          example: "John Doe"
    SignIn:
      type: object
      properties:
        email:
          type: "string"
          format: "email"
          example: "johndoe@example.com"
        username:
          type: "string"
          pattern: "^[A-Za-z0-9]*$"
          example: "asjfhasfjhasf"
        password:
          type: "string"
          pattern: "^[A-Za-z0-9!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?]{8,}$"
          example: "p@ssw0rd123!"
        playerName:
          type: "string"
          pattern: "^[A-Za-zА-Яа-я0-9]+(?: [A-Za-zА-Яа-я0-9]+)*$"
          example: "John Doe"
    SignOut:
      type: object
      properties:
        id:
          type: "string"
        refreshToken:
          type: "string"
    PlayerResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        seat:
          type: "integer"
          format: "int32"
        name:
          type: "string"
        role:
          type: "string"
          enum:
            - "CITIZEN"
            - "MAFIOSO"
            - "DON"
            - "DETECTIVE"
        status:
          type: "string"
          enum:
            - "INGAME"
            - "ALIVE"
            - "DEAD"
            - "REMOVED"
            - "OPW"
        inGame:
          type: "boolean"
        bonusPoints:
          type: "integer"
          format: "int32"
...
    DayVoteResponse:
      type: "object"
      properties:
        dayNumber:
          type: "integer"
          format: "int32"
        target:
          $ref: "#/components/schemas/PlayerResponse"
        player:
          $ref: "#/components/schemas/PlayerResponse"
    DayStageResponse:
      type: "object"
      properties:
        type:
          type: "string"
          enum:
            - "DAY"
        dayNumber:
          type: "integer"
          format: "int32"
        candidates:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        votes:
          type: "array"
          items:
            $ref: "#/components/schemas/DayVoteResponse"
    NightStageResponse:
      type: "object"
      properties:
        dayNumber:
          type: "integer"
          format: "int32"
        mafiaShot:
          $ref: "#/components/schemas/PlayerResponse"
        detectiveCheck:
          $ref: "#/components/schemas/PlayerResponse"
        donCheck:
          $ref: "#/components/schemas/PlayerResponse"
    GameResponse:
      type: "object"
      properties:
        id:
          type: "string"
        date:
          type: "string"
        players:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        currentStage:
          anyOf:
            - title: "DayStageResponse"
              $ref: "#/components/schemas/DayStageResponse"
            - title: "NightStageResponse"
              $ref: "#/components/schemas/NightStageResponse"
        nominates:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
        votes:
          type: "array"
          items:
            $ref: "#/components/schemas/DayVoteResponse"
        mafiaShot:
          $ref: "#/components/schemas/PlayerResponse"
        detectiveCheck:
          $ref: "#/components/schemas/PlayerResponse"
        donCheck:
          $ref: "#/components/schemas/PlayerResponse"
        playersEliminated:
          type: "array"
          items:
            $ref: "#/components/schemas/PlayerResponse"
...
    NoContent:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "int32"
          example: 204
      required:
        - "status"
    Error:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "int32"
          example: 400
        message:
          type: "string"
          description: "A description of the error."
          example: "Invalid ID"
      required:
        - "status"
        - "message"
  responses:
    BadRequest:
      description: "Malformed data or otherwise invalid request"
      schema:
        $ref: "#/components/schemas/Error"
    NotFound:
      description: "The requested component was not found"
      schema:
        $ref: "#/components/schemas/Error"
    InternalServerError:
      description: "An internal server error occurred"
      schema:
        $ref: "#/components/schemas/Error"
    Forbidden:
      description: "The user does not have the required permissions to access this resource"
      schema:
        $ref: "#/components/schemas/Error"
    Unauthorized:
      description: "The user is not authenticated"
      schema:
        $ref: "#/components/schemas/Error"